src/requirements.txt
flask==3.0.2
python-dotenv==1.0.1
gunicorn==21.2.0
click==8.1.7